Moved filedlg fixture to its own file
